requirements.txt
aiohttp==3.6.2
Pillow==6.2.1